+++ python/mxnet/contrib/onnx/mx2onnx/LICENSE
