+++ python/mxnet/contrib/onnx/mx2onnx/__init__.py
@@ -21,4 +21,4 @@ from __future__ import absolute_import
 
 from . import export_model
 from . import export_onnx
-from . import op_translations
+from . import _op_translations
