+++ tests/python-pytest/onnx/import/mxnet_backend_rep.py
@@ -17,7 +17,6 @@
 
 # coding: utf-8
 """MXNet backend rep for onnx test infrastructure"""
-import numpy as np
 try:
     from onnx.backend.base import BackendRep
 except ImportError:
